🚚 api: move controller to session-management context
