+++ api/lib/application/certification-candidates/certification-candidates-controller.js
@@ -1,15 +1,5 @@
 import { usecases } from '../../domain/usecases/index.js';
 
-const authorizeToResume = async function (request, h) {
-  const certificationCandidateId = request.params.id;
-
-  await usecases.authorizeCertificationCandidateToResume({
-    certificationCandidateId,
-  });
-
-  return h.response().code(204);
-};
-
 const endAssessmentBySupervisor = async function (request) {
   const certificationCandidateId = request.params.id;
 
@@ -19,7 +9,6 @@ const endAssessmentBySupervisor = async function (request) {
 };
 
 const certificationCandidatesController = {
-  authorizeToResume,
   endAssessmentBySupervisor,
 };
